requirements.txt
fastapi
uvicorn
gradio
python-multipart
langchain-google-genai
pymupdf
python-dotenv
pydantic
